+++ DutWrapper/Model/Account/AccountInformation.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DutWrapper.Model.Enums;
+using System;
 
 namespace DutWrapper.Model.Account
 {
@@ -20,14 +21,54 @@ namespace DutWrapper.Model.Account
         public DateTime DateOfBirth { get; set; }
 
         /// <summary>
+        /// Student date place
+        /// </summary>
+        public string? DatePlace { get; set; }
+
+        /// <summary>
         /// Student gender
         /// </summary>
         public Gender Gender { get; set; } = Gender.Unknown;
 
         /// <summary>
-        /// Student identity ID
+        /// Student ethnicity
         /// </summary>
-        public string? IdentityID { get; set; } = null;
+        public string? Ethnicity { get; set; } = null;
+
+        /// <summary>
+        /// Student nationality
+        /// </summary>
+        public string? Nationality { get; set; } = null;
+
+        /// <summary>
+        /// Student religion
+        /// </summary>
+        public string? Religion { get; set; } = null;
+
+        /// <summary>
+        /// Student national card ID
+        /// </summary>
+        public string? NationalCardID { get; set; } = null;
+
+        /// <summary>
+        /// Student national card issue date
+        /// </summary>
+        public DateTime NationalCardIssueDate { get; set; }
+
+        /// <summary>
+        /// Student national card issue place
+        /// </summary>
+        public string? NationalCardIssuePlace { get; set; } = null;
+
+        /// <summary>
+        /// Student citizen card ID (replace for national card)
+        /// </summary>
+        public string? CitizenCardID { get; set; } = null;
+
+        /// <summary>
+        /// Student citizen card issue date.
+        /// </summary>
+        public DateTime CitizenCardIssueDate { get; set; }
 
         /// <summary>
         /// Student bank information
@@ -37,7 +78,32 @@ namespace DutWrapper.Model.Account
         /// <summary>
         /// Student health insurance ID
         /// </summary>
-        public string? HIID { get; set; } = null;
+        public string? HealthInsuranceID { get; set; } = null;
+
+        /// <summary>
+        /// Student health insurance expiration date
+        /// </summary>
+        public DateTime HealthInsuranceExpirationDate { get; set; }
+
+        /// <summary>
+        /// Student specialization
+        /// </summary>
+        public string? Specialization { get; set; } = null;
+
+        /// <summary>
+        /// Student main training program plan
+        /// </summary>
+        public string? TrainingProgramPlan { get; set; } = null;
+
+        /// <summary>
+        /// Student second training program plan (if available)
+        /// </summary>
+        public string? TrainingProgramPlan2 { get; set; } = null;
+
+        /// <summary>
+        /// Student email (school provided, ex. Microsoft, Google,...)
+        /// </summary>
+        public string? SchoolEmail { get; set; } = null;
 
         /// <summary>
         /// Student email
@@ -50,9 +116,9 @@ namespace DutWrapper.Model.Account
         public string? PhoneNumber { get; set; }
 
         /// <summary>
-        /// Student account email for education
+        /// Student facebook link
         /// </summary>
-        public string? EducationEmail { get; set; } = null;
+        public string? FacebookLink { get; set; } = null;
 
         /// <summary>
         /// Student class name (ex. 20R12)